+++ app/src/main/kotlin/com/emm/justchill/hh/data/transaction/TransactionInsert.kt
@@ -3,7 +3,7 @@ package com.emm.justchill.hh.data.transaction
 data class TransactionInsert(
     val id: String? = null,
     val type: String,
-    val amount: Long = 0L,
+    val amount: Double = 0.0,
     val description: String,
     val date: Long,
 )
